requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
confusable-homoglyphs==3.2.0
cryptography==2.7
defusedxml==0.5.0
Django==2.2.3
django-allauth==0.39.1
django-auth==0.1.9
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-registration==3.0
django-rest-auth==0.9.3
django-webpack-loader==0.6.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
get==2019.4.13
gistory==0.44
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
Pillow==6.0.0
post==2019.4.13
public==2019.4.13
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
python3-openid==3.1.0
pytz==2018.9
query-string==2019.4.13
request==2019.4.13
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.6.0
wrapt==1.11.1
